Miscellaneous: New Features, Refactors, Bindings, API, Typing (Part 7)
Renamed the contact interface to neighbors (not a good name). Continued
the simplification of the Luni class, which unfortunately, used to be a super class.
I moved the tokenization and the entire selection parsing code to a
separate directory. All of this will be replaced soon with the new
implementation. For know, it's there to avoid breaking tests.

Many other smaller changes.
