+++ cpp/include/selections/nodes.hpp
