+++ cpp/include/selections/operators.hpp
