+++ cpp/include/selections/visitor.hpp
