+++ cpp/src/convert.cpp
@@ -1,4 +1,5 @@
 #include "convert.hpp"
+#include "spdlog/spdlog.h"
 #include <rdkit/GraphMol/MonomerInfo.h>
 #include <unordered_map>
 
@@ -112,53 +113,58 @@ RWMol filter_with_conf(RWMol &mol, std::vector<int> &indices) {
   return new_mol;
 }
 
-RWMol filter_with_bonds(const RWMol &mol, const std::vector<int> &indices) {
-  RWMol new_mol;
+
+RWMol filter_with_bonds(const RWMol &mol, std::vector<int> &indices) { // indices do not need to be sorted (they are sorted in the function)
+  RWMol filtered_mol;
   const Conformer &conf = mol.getConformer();
-  auto *new_conf = new Conformer();
+  auto *filtered_conf = new Conformer();
 
-  // for faster index lookup
-  std::unordered_map<int, int> old_to_new_index;
+  std::unordered_map<int, int> atom_index_map;
 
-  // sanity check
-  auto max_idx = std::max_element(indices.begin(), indices.end());
-  if (max_idx == indices.end() || *max_idx >= mol.getNumAtoms()) {
-    new_mol.addConformer(new_conf, true);
-    new_mol.updatePropertyCache(false);
-    return new_mol;
+  if (indices.empty()) {
+    filtered_mol.addConformer(filtered_conf, true);
+    filtered_mol.updatePropertyCache(false);
+    return filtered_mol;
   }
 
+  std::sort(indices.begin(), indices.end());
+
+  if (indices.back() >= mol.getNumAtoms()) {
+    std::string err_msg = fmt::format("Invalid atom index {} in the provided indices. Max index: {}", 
+                                      indices.back(), mol.getNumAtoms() - 1);
+    spdlog::critical("Error: {}", err_msg);
+    throw std::runtime_error(err_msg);
+  }
+
+  // copy atoms and positions.
   for (size_t i = 0; i < indices.size(); ++i) {
-    int idx = indices[i];
-    const RDKit::Atom *atom = mol.getAtomWithIdx(idx);
-    int new_idx = new_mol.addAtom(new RDKit::Atom(*atom), false, true);
-    old_to_new_index[idx] = new_idx;
-    new_conf->setAtomPos(new_idx, conf.getAtomPos(idx));
+    int atom_idx = indices[i];
+    const RDKit::Atom *atom = mol.getAtomWithIdx(atom_idx);
+    int filtered_idx = filtered_mol.addAtom(new RDKit::Atom(*atom), false, true);
+    atom_index_map[atom_idx] = filtered_idx;
+    filtered_conf->setAtomPos(filtered_idx, conf.getAtomPos(atom_idx));
   }
 
-  for (int old_idx : indices) {
-    for (const auto &bond : mol.atomBonds(mol.getAtomWithIdx(old_idx))) {
-      int begin_idx = bond->getBeginAtomIdx();
-      int end_idx = bond->getEndAtomIdx();
+  // add bonds only if both atoms are in the selected indices.
+  for (const auto &bond : mol.bonds()) {
+    int start_idx = bond->getBeginAtomIdx();
+    int end_idx   = bond->getEndAtomIdx();
 
-      auto begin_it = old_to_new_index.find(begin_idx);
-      auto end_it = old_to_new_index.find(end_idx);
-
-      if (begin_it != old_to_new_index.end() && end_it != old_to_new_index.end()) {
-        int new_begin_idx = begin_it->second;
-        int new_end_idx = end_it->second;
-        if (new_begin_idx > new_end_idx) std::swap(new_begin_idx, new_end_idx);
-
-        if (!new_mol.getBondBetweenAtoms(new_begin_idx, new_end_idx)) {
-          new_mol.addBond(new_begin_idx, new_end_idx, bond->getBondType());
-        }
-      }
+    if (atom_index_map.find(start_idx) != atom_index_map.end() && atom_index_map.find(end_idx) != atom_index_map.end()) {
+      int new_start_idx = atom_index_map[start_idx];
+      int new_end_idx   = atom_index_map[end_idx];
+      filtered_mol.addBond(new_start_idx, new_end_idx, bond->getBondType());
     }
   }
 
-  new_mol.addConformer(new_conf, true);
-  new_mol.updatePropertyCache(false);
-  return new_mol;
+  filtered_mol.addConformer(filtered_conf, true);
+  filtered_mol.updatePropertyCache(false);
+
+
+  spdlog::warn("Filtered molecule has {} atoms and {} bonds", filtered_mol.getNumAtoms(), filtered_mol.getNumBonds());
+
+  return filtered_mol;
 }
+
 
 } // namespace lahuta
